new try

+++ treemap.c
@@ -80,7 +80,7 @@ Pair * searchTreeMap(TreeMap * tree, void* key) {
             tree->current = aux;
             return aux->pair;
         }
-        if (key > aux->pair->key && aux->right != NULL)
+        if ((int)key > (int)aux->pair->key && aux->right != NULL)
             aux = aux->right;
         else if (key < aux->pair->key && aux->left != NULL)
             aux = aux->left;
